refactor: extract ci-metrics into separate server
Move all new metrics, billing, and dashboard code from rkapp into
ci3/ci-metrics/ as a self-contained Flask server with its own Redis
connection, SQLite DB, Dockerfile, and background threads.

rkapp (rk.py) stripped to original CI log viewer + thin proxy that
forwards /api/* and dashboard routes to ci-metrics. Starts ci-metrics
as a gunicorn subprocess on startup.

Remove namespace→user attribution heuristic (branch_to_ns_prefix);
GCP costs reported as flat total pending BigQuery billing export.
