+++ ci3/ci-metrics/requirements.txt
@@ -1,7 +1,8 @@
 flask
 gunicorn
 redis
-ansi2html
 Flask-Compress
-requests
 Flask-HTTPAuth
+requests
+google-cloud-bigquery
+boto3
