+++ ci3/ci-metrics/views/ci-insights.html
