+++ ci3/dashboard/rk.py
@@ -1,7 +1,6 @@
 from flask import Flask, render_template_string, request, Response, redirect
 from flask_compress import Compress
 from flask_httpauth import HTTPBasicAuth
-from datetime import datetime, timedelta
 import gzip
 import json
 import os
@@ -19,34 +18,27 @@ from rk_core import (
     YELLOW, BLUE, GREEN, RED, PURPLE, BOLD, RESET,
     hyperlink, r, get_section_data, get_list_as_string
 )
-from billing import (
-    get_billing_files_in_range,
-    aggregate_billing_weekly, aggregate_billing_monthly,
-    serve_billing_dashboard,
-)
-import rk_db
-import rk_metrics
-import billing.aws as rk_aws_costs
-import rk_github
-
 LOGS_DISK_PATH = os.getenv('LOGS_DISK_PATH', '/logs-disk')
 DASHBOARD_PASSWORD = os.getenv('DASHBOARD_PASSWORD', 'password')
+CI_METRICS_PORT = int(os.getenv('CI_METRICS_PORT', '8081'))
+CI_METRICS_URL = os.getenv('CI_METRICS_URL', f'http://localhost:{CI_METRICS_PORT}')
+
 app = Flask(__name__)
 Compress(app)
 auth = HTTPBasicAuth()
 
-def _init_metrics():
-    """Initialize SQLite for test events, start test listener and CI run sync."""
-    try:
-        rk_db.get_db()
-        rk_metrics.start_test_listener(r)
-        rk_metrics.start_ci_run_sync(r)
-        print("[rk.py] Test event listener and CI run sync started")
-    except Exception as e:
-        print(f"[rk.py] Warning: metrics startup failed: {e}")
-
-# Start in background to avoid blocking Flask startup
-threading.Thread(target=_init_metrics, daemon=True, name='metrics-init').start()
+# Start the ci-metrics server as a subprocess
+_ci_metrics_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), '..', 'ci-metrics')
+if os.path.isdir(_ci_metrics_dir):
+    _ci_metrics_env = {**os.environ, 'CI_METRICS_PORT': str(CI_METRICS_PORT)}
+    subprocess.Popen(
+        ['gunicorn', '-w', '4', '-b', f'0.0.0.0:{CI_METRICS_PORT}', 'app:app'],
+        cwd=_ci_metrics_dir,
+        env=_ci_metrics_env,
+        stdout=subprocess.DEVNULL,
+        stderr=subprocess.DEVNULL,
+    )
+    print(f"[rk.py] ci-metrics server started on port {CI_METRICS_PORT}")
 
 def read_from_disk(key):
     """Read log from disk as fallback when Redis key not found."""
@@ -169,11 +161,11 @@ def root() -> str:
         f"{hyperlink('/chonk-breakdowns', 'chonk breakdowns')}\n"
         f"{RESET}"
         f"\n"
-        f"Billing:\n"
+        f"CI Metrics:\n"
         f"\n{YELLOW}"
-        f"{hyperlink('/cost-overview', 'cost overview (AWS + GCP)')}\n"
-        f"{hyperlink('/namespace-billing', 'namespace billing')}\n"
-        f"{hyperlink('/ci-insights', 'ci insights')}\n"
+        f"{hyperlink(CI_METRICS_URL + '/cost-overview', 'cost overview (AWS + GCP)')}\n"
+        f"{hyperlink(CI_METRICS_URL + '/namespace-billing', 'namespace billing')}\n"
+        f"{hyperlink(CI_METRICS_URL + '/ci-insights', 'ci insights')}\n"
         f"{RESET}"
     )
 
@@ -518,750 +510,31 @@ def trigger_grind():
     return redirect(f'/{run_id}')
 
 
+# ---- Thin proxy to ci-metrics server ----
+
 @app.route('/namespace-billing')
-@auth.login_required
-def namespace_billing():
-    html = serve_billing_dashboard()
-    if html:
-        return html
-    return "Billing dashboard not found", 404
-
-@app.route('/api/billing/data')
-@auth.login_required
-def billing_data():
-    date_from_str = request.args.get('from')
-    date_to_str = request.args.get('to')
-    granularity = request.args.get('granularity', 'daily')
-
-    if not date_from_str or not date_to_str:
-        return Response(json.dumps({'error': 'from and to date params required (YYYY-MM-DD)'}),
-                        mimetype='application/json', status=400)
-    try:
-        date_from = datetime.strptime(date_from_str, '%Y-%m-%d')
-        date_to = datetime.strptime(date_to_str, '%Y-%m-%d')
-    except ValueError:
-        return Response(json.dumps({'error': 'Invalid date format, use YYYY-MM-DD'}),
-                        mimetype='application/json', status=400)
-
-    daily_data = get_billing_files_in_range(date_from, date_to)
-
-    # Filter out namespaces costing less than $1 total across the range
-    ns_totals = {}
-    for entry in daily_data:
-        for ns, ns_data in entry.get('namespaces', {}).items():
-            ns_totals[ns] = ns_totals.get(ns, 0) + ns_data.get('total', 0)
-    cheap_ns = {ns for ns, total in ns_totals.items() if total < 1.0}
-    if cheap_ns:
-        for entry in daily_data:
-            entry['namespaces'] = {ns: d for ns, d in entry.get('namespaces', {}).items()
-                                   if ns not in cheap_ns}
-
-    if granularity == 'weekly':
-        result = aggregate_billing_weekly(daily_data)
-    elif granularity == 'monthly':
-        result = aggregate_billing_monthly(daily_data)
-    else:
-        result = daily_data
-
-    return Response(json.dumps(result), mimetype='application/json')
-
-
-def _aggregate_dates(by_date_list, granularity, sum_fields, avg_fields=None):
-    """Aggregate a list of {date, ...} dicts by weekly/monthly granularity."""
-    if granularity == 'daily' or not by_date_list:
-        return by_date_list
-
-    buckets = {}
-    for entry in by_date_list:
-        d = datetime.strptime(entry['date'], '%Y-%m-%d')
-        if granularity == 'weekly':
-            key = (d - timedelta(days=d.weekday())).strftime('%Y-%m-%d')
-        else:  # monthly
-            key = d.strftime('%Y-%m') + '-01'
-
-        if key not in buckets:
-            buckets[key] = {'date': key}
-            for f in sum_fields:
-                buckets[key][f] = 0
-            if avg_fields:
-                for f in avg_fields:
-                    buckets[key][f'_avg_sum_{f}'] = 0
-                    buckets[key][f'_avg_cnt_{f}'] = 0
-
-        for f in sum_fields:
-            buckets[key][f] += entry.get(f) or 0
-        if avg_fields:
-            for f in avg_fields:
-                val = entry.get(f)
-                if val is not None:
-                    buckets[key][f'_avg_sum_{f}'] += val
-                    buckets[key][f'_avg_cnt_{f}'] += 1
-
-    result = []
-    for key in sorted(buckets):
-        b = buckets[key]
-        out = {'date': b['date']}
-        for f in sum_fields:
-            out[f] = round(b[f], 2) if isinstance(b[f], float) else b[f]
-        if avg_fields:
-            for f in avg_fields:
-                cnt = b[f'_avg_cnt_{f}']
-                out[f] = round(b[f'_avg_sum_{f}'] / cnt, 1) if cnt else None
-        result.append(out)
-
-    return result
-
-
-@app.route('/api/ci/runs')
-@auth.login_required
-def api_ci_runs():
-    """API endpoint: list CI runs from Redis with optional filters."""
-    date_from = request.args.get('from', '')
-    date_to = request.args.get('to', '')
-    status_filter = request.args.get('status', '')
-    author = request.args.get('author', '')
-    dashboard = request.args.get('dashboard', '')
-    limit = min(int(request.args.get('limit', 100)), 1000)
-    offset = int(request.args.get('offset', 0))
-
-    ts_from = int(datetime.strptime(date_from, '%Y-%m-%d').timestamp() * 1000) if date_from else None
-    ts_to = int((datetime.strptime(date_to, '%Y-%m-%d') + timedelta(days=1)).timestamp() * 1000) if date_to else None
-
-    runs = rk_metrics.get_ci_runs(r, ts_from, ts_to)
-
-    if status_filter:
-        runs = [run for run in runs if run.get('status') == status_filter]
-    if author:
-        runs = [run for run in runs if run.get('author') == author]
-    if dashboard:
-        runs = [run for run in runs if run.get('dashboard') == dashboard]
-
-    runs.sort(key=lambda x: x.get('timestamp', 0), reverse=True)
-    runs = runs[offset:offset + limit]
-
-    return Response(json.dumps(runs), mimetype='application/json')
-
-
-@app.route('/api/ci/stats')
-@auth.login_required
-def api_ci_stats():
-    """API endpoint: CI run statistics summary (last 7 days)."""
-    ts_from = int((datetime.now() - timedelta(days=7)).timestamp() * 1000)
-    runs = rk_metrics.get_ci_runs(r, ts_from)
-
-    total = len(runs)
-    passed = sum(1 for run in runs if run.get('status') == 'PASSED')
-    failed = sum(1 for run in runs if run.get('status') == 'FAILED')
-    costs = [run['cost_usd'] for run in runs if run.get('cost_usd') is not None]
-    durations = []
-    for run in runs:
-        complete = run.get('complete')
-        ts = run.get('timestamp')
-        if complete and ts:
-            durations.append((complete - ts) / 60000.0)
-
-    return Response(json.dumps({
-        'total_runs': total,
-        'passed': passed,
-        'failed': failed,
-        'total_cost': round(sum(costs), 2) if costs else None,
-        'avg_duration_mins': round(sum(durations) / len(durations), 1) if durations else None,
-    }), mimetype='application/json')
-
-
-# ---- Cost endpoints ----
-
-@app.route('/api/costs/overview')
-@auth.login_required
-def api_costs_overview():
-    date_from = request.args.get('from', (datetime.now() - timedelta(days=30)).strftime('%Y-%m-%d'))
-    date_to = request.args.get('to', datetime.now().strftime('%Y-%m-%d'))
-    granularity = request.args.get('granularity', 'daily')
-    result = rk_aws_costs.get_costs_overview(date_from, date_to)
-    if granularity != 'daily' and result.get('by_date'):
-        buckets = {}
-        for entry in result['by_date']:
-            d = datetime.strptime(entry['date'], '%Y-%m-%d')
-            if granularity == 'weekly':
-                key = (d - timedelta(days=d.weekday())).strftime('%Y-%m-%d')
-            else:
-                key = d.strftime('%Y-%m') + '-01'
-            if key not in buckets:
-                buckets[key] = {'date': key, 'aws': {}, 'gcp': {}, 'aws_total': 0, 'gcp_total': 0}
-            for cat, amt in entry.get('aws', {}).items():
-                buckets[key]['aws'][cat] = buckets[key]['aws'].get(cat, 0) + amt
-            for cat, amt in entry.get('gcp', {}).items():
-                buckets[key]['gcp'][cat] = buckets[key]['gcp'].get(cat, 0) + amt
-            buckets[key]['aws_total'] += entry.get('aws_total', 0)
-            buckets[key]['gcp_total'] += entry.get('gcp_total', 0)
-        result['by_date'] = sorted(buckets.values(), key=lambda x: x['date'])
-    return Response(json.dumps(result), mimetype='application/json')
-
-
-@app.route('/api/costs/details')
-@auth.login_required
-def api_costs_details():
-    """Per-resource (USAGE_TYPE) cost breakdown — shows individual instance types and RI fees."""
-    date_from = request.args.get('from', (datetime.now() - timedelta(days=30)).strftime('%Y-%m-%d'))
-    date_to = request.args.get('to', datetime.now().strftime('%Y-%m-%d'))
-
-    rows = rk_aws_costs.get_aws_cost_details(date_from, date_to)
-
-    # Aggregate: usage_type -> { total, by_date: {date: amount}, is_ri }
-    usage_map = {}
-    for row in rows:
-        ut = row['usage_type']
-        if ut not in usage_map:
-            usage_map[ut] = {
-                'usage_type': ut,
-                'service': row['service'],
-                'category': row['category'],
-                'total': 0,
-                'by_date': {},
-                'is_ri': 'HeavyUsage' in ut,
-            }
-        usage_map[ut]['total'] += row['amount_usd']
-        d = row['date']
-        usage_map[ut]['by_date'][d] = usage_map[ut]['by_date'].get(d, 0) + row['amount_usd']
-
-    # Sort by total cost descending
-    items = sorted(usage_map.values(), key=lambda x: -x['total'])
-
-    # Round totals
-    for item in items:
-        item['total'] = round(item['total'], 2)
-        item['by_date'] = {d: round(v, 4) for d, v in sorted(item['by_date'].items())}
-
-    # Collect all dates for the table header
-    all_dates = sorted({row['date'] for row in rows})
-
-    # RI summary
-    ri_items = [i for i in items if i['is_ri']]
-    ri_total = round(sum(i['total'] for i in ri_items), 2)
-
-    return Response(json.dumps({
-        'items': items,
-        'dates': all_dates,
-        'ri_total': ri_total,
-        'grand_total': round(sum(i['total'] for i in items), 2),
-    }), mimetype='application/json')
-
-
-@app.route('/api/costs/attribution')
-@auth.login_required
-def api_costs_attribution():
-    """CI cost attribution by user, branch, instance — from Redis CI run data + GKE namespace billing."""
-    date_from = request.args.get('from', (datetime.now() - timedelta(days=30)).strftime('%Y-%m-%d'))
-    date_to = request.args.get('to', datetime.now().strftime('%Y-%m-%d'))
-    ts_from = int(datetime.strptime(date_from, '%Y-%m-%d').timestamp() * 1000)
-    ts_to = int((datetime.strptime(date_to, '%Y-%m-%d') + timedelta(days=1)).timestamp() * 1000)
-
-    runs = rk_metrics.get_ci_runs(r, ts_from, ts_to)
-    runs_with_cost = [run for run in runs if run.get('cost_usd') is not None]
-
-    # Enrich merge queue runs with PR author from GitHub
-    pr_numbers = {run.get('pr_number') for run in runs_with_cost if run.get('pr_number')}
-    pr_authors = rk_github.batch_get_pr_authors(pr_numbers)
-
-    granularity = request.args.get('granularity', 'daily')
-
-    # Build per-instance records with decoded branch info
-    instances = []
-    by_user = {}
-    by_branch = {}
-    by_type = {}  # merge-queue, pr, nightly, etc.
-    by_date_type = {}  # date -> type -> cost (time series)
-
-    for run in runs_with_cost:
-        info = rk_aws_costs.decode_branch_info(run)
-        cost = run['cost_usd']
-        date = rk_metrics._ts_to_date(run.get('timestamp', 0))
-
-        # Resolve author: for merge queue, prefer GitHub PR author over git commit author
-        author = info['author']
-        prn = info['pr_number']
-        if prn and int(prn) in pr_authors:
-            author = pr_authors[int(prn)]['author']
-
-        inst_type = run.get('instance_type', 'unknown')
-        vcpus = run.get('instance_vcpus')
-        if inst_type == 'unknown' and vcpus:
-            inst_type = f'{vcpus}vcpu'
-
-        instances.append({
-            'instance_name': info['instance_name'],
-            'date': date,
-            'cost_usd': cost,
-            'author': author,
-            'branch': info['branch'],
-            'pr_number': prn,
-            'type': info['type'],
-            'instance_type': inst_type,
-            'spot': run.get('spot', False),
-            'job_id': run.get('job_id', ''),
-            'duration_mins': round((run.get('complete', 0) - run.get('timestamp', 0)) / 60000, 1) if run.get('complete') else None,
-        })
-
-        # Aggregate by user
-        if author not in by_user:
-            by_user[author] = {'aws_cost': 0, 'gcp_cost': 0, 'runs': 0, 'by_date': {}}
-        by_user[author]['aws_cost'] += cost
-        by_user[author]['runs'] += 1
-        by_user[author]['by_date'][date] = by_user[author]['by_date'].get(date, 0) + cost
-
-        # Aggregate by branch
-        branch_key = info['branch'] or info['type']
-        if branch_key not in by_branch:
-            by_branch[branch_key] = {'cost': 0, 'runs': 0, 'type': info['type'], 'author': author}
-        by_branch[branch_key]['cost'] += cost
-        by_branch[branch_key]['runs'] += 1
-
-        # Aggregate by type
-        rt = info['type']
-        if rt not in by_type:
-            by_type[rt] = {'cost': 0, 'runs': 0}
-        by_type[rt]['cost'] += cost
-        by_type[rt]['runs'] += 1
-
-        # Time series: cost by date and type
-        if date not in by_date_type:
-            by_date_type[date] = {}
-        by_date_type[date][rt] = by_date_type[date].get(rt, 0) + cost
-
-    # Add GKE namespace costs attributed to users (only if we can match to a GitHub actor).
-    # Namespace naming: pr-$(echo "$branch" | sed 's/[^a-z0-9-]/-/g' | cut -c1-20 | sed 's/-*$//')
-    # So we match namespaces to CI run branch names by reversing the sanitization.
-    gcp_unattributed = 0
-    try:
-        from billing.gcp import get_billing_files_in_range
-        gcp_data = get_billing_files_in_range(
-            datetime.strptime(date_from, '%Y-%m-%d'),
-            datetime.strptime(date_to, '%Y-%m-%d'),
-        )
-
-        # Build namespace -> author mapping from CI runs (branch name -> sanitized namespace prefix)
-        import re
-        def branch_to_ns_prefix(branch):
-            """Replicate: echo "$branch" | sed 's/[^a-z0-9-]/-/g' | cut -c1-20 | sed 's/-*$//'"""
-            sanitized = re.sub(r'[^a-z0-9-]', '-', branch.lower())[:20].rstrip('-')
-            return f'pr-{sanitized}'
-
-        # Map sanitized prefixes to authors from CI runs
-        ns_prefix_to_author = {}
-        for run in runs_with_cost:
-            branch = run.get('name', '')
-            if not branch or '(queue)' in branch:
-                continue
-            prefix = branch_to_ns_prefix(branch)
-            author = run.get('author', 'unknown')
-            prn = run.get('pr_number')
-            if prn and int(prn) in pr_authors:
-                author = pr_authors[int(prn)]['author']
-            ns_prefix_to_author[prefix] = author
-
-        for entry in gcp_data:
-            for ns, ns_data in entry.get('namespaces', {}).items():
-                ns_cost = ns_data.get('total', 0)
-                if ns_cost < 0.01:
-                    continue
-                # Try to match namespace to a known GitHub actor
-                ns_author = None
-                if ns.startswith('pr-'):
-                    # Strip scenario test suffix (-1, -2, etc.) for matching
-                    ns_base = re.sub(r'-\d+$', '', ns)
-                    ns_author = ns_prefix_to_author.get(ns) or ns_prefix_to_author.get(ns_base)
-
-                if ns_author:
-                    if ns_author not in by_user:
-                        by_user[ns_author] = {'aws_cost': 0, 'gcp_cost': 0, 'runs': 0, 'by_date': {}}
-                    by_user[ns_author]['gcp_cost'] += ns_cost
-                else:
-                    gcp_unattributed += ns_cost
-    except Exception as e:
-        print(f"[attribution] GKE enrichment error: {e}")
-
-    # Sort and format
-    user_list = [{'author': a, 'aws_cost': round(v['aws_cost'], 2), 'gcp_cost': round(v['gcp_cost'], 2),
-                  'total_cost': round(v['aws_cost'] + v['gcp_cost'], 2), 'runs': v['runs'],
-                  'by_date': {d: round(c, 2) for d, c in sorted(v['by_date'].items())}}
-                 for a, v in sorted(by_user.items(), key=lambda x: -(x[1]['aws_cost'] + x[1]['gcp_cost']))]
-
-    branch_list = [{'branch': b, 'cost': round(v['cost'], 2), 'runs': v['runs'],
-                     'type': v['type'], 'author': v['author']}
-                    for b, v in sorted(by_branch.items(), key=lambda x: -x[1]['cost'])[:100]]
-
-    type_list = [{'type': t, 'cost': round(v['cost'], 2), 'runs': v['runs']}
-                 for t, v in sorted(by_type.items(), key=lambda x: -x[1]['cost'])]
-
-    instances.sort(key=lambda x: -(x['cost_usd'] or 0))
-
-    # Build time series: [{date, total, runs, merge-queue, pr, nightly, ...}]
-    all_types = sorted(by_type.keys())
-    by_date_list = []
-    for date in sorted(by_date_type):
-        entry = {'date': date, 'total': 0, 'runs': 0}
-        for rt in all_types:
-            entry[rt] = round(by_date_type[date].get(rt, 0), 2)
-            entry['total'] += by_date_type[date].get(rt, 0)
-        entry['total'] = round(entry['total'], 2)
-        # Count runs for this date
-        entry['runs'] = sum(1 for inst in instances if inst['date'] == date)
-        by_date_list.append(entry)
-
-    by_date_list = _aggregate_dates(by_date_list, granularity,
-                                     sum_fields=['total', 'runs'] + all_types)
-
-    total_aws = sum(u['aws_cost'] for u in user_list)
-    total_gcp = sum(u['gcp_cost'] for u in user_list) + gcp_unattributed
-
-    return Response(json.dumps({
-        'by_user': user_list,
-        'by_branch': branch_list,
-        'by_type': type_list,
-        'by_date': by_date_list,
-        'run_types': all_types,
-        'instances': instances[:500],  # cap to avoid huge payloads
-        'totals': {'aws': round(total_aws, 2), 'gcp': round(total_gcp, 2),
-                   'gcp_unattributed': round(gcp_unattributed, 2),
-                   'combined': round(total_aws + total_gcp, 2)},
-    }), mimetype='application/json')
-
-
-@app.route('/api/costs/runners')
-@auth.login_required
-def api_costs_runners():
-    date_from = request.args.get('from', (datetime.now() - timedelta(days=30)).strftime('%Y-%m-%d'))
-    date_to = request.args.get('to', datetime.now().strftime('%Y-%m-%d'))
-    granularity = request.args.get('granularity', 'daily')
-    dashboard = request.args.get('dashboard', '')
-    ts_from = int(datetime.strptime(date_from, '%Y-%m-%d').timestamp() * 1000)
-    ts_to = int((datetime.strptime(date_to, '%Y-%m-%d') + timedelta(days=1)).timestamp() * 1000)
-
-    runs = rk_metrics.get_ci_runs(r, ts_from, ts_to)
-    runs_with_cost = [run for run in runs if run.get('cost_usd') is not None]
-    if dashboard:
-        runs_with_cost = [run for run in runs_with_cost if run.get('dashboard') == dashboard]
-
-    # By date
-    by_date_map = {}
-    for run in runs_with_cost:
-        date = rk_metrics._ts_to_date(run.get('timestamp', 0))
-        if date not in by_date_map:
-            by_date_map[date] = {'spot_cost': 0, 'ondemand_cost': 0, 'total': 0}
-        cost = run['cost_usd']
-        if run.get('spot'):
-            by_date_map[date]['spot_cost'] += cost
-        else:
-            by_date_map[date]['ondemand_cost'] += cost
-        by_date_map[date]['total'] += cost
-
-    by_date = [{'date': date, 'spot_cost': round(d['spot_cost'], 2),
-                'ondemand_cost': round(d['ondemand_cost'], 2), 'total': round(d['total'], 2),
-                'spot_pct': round(100.0 * d['spot_cost'] / max(d['total'], 0.01), 1)}
-               for date, d in sorted(by_date_map.items())]
-
-    by_date = _aggregate_dates(by_date, granularity,
-                               sum_fields=['spot_cost', 'ondemand_cost', 'total'])
-    for d in by_date:
-        d['spot_pct'] = round(100.0 * d['spot_cost'] / max(d['total'], 0.01), 1)
-
-    # By instance type
-    by_instance_map = {}
-    for run in runs_with_cost:
-        inst = run.get('instance_type', 'unknown')
-        if inst not in by_instance_map:
-            by_instance_map[inst] = {'cost': 0, 'runs': 0}
-        by_instance_map[inst]['cost'] += run['cost_usd']
-        by_instance_map[inst]['runs'] += 1
-    by_instance = [{'instance_type': k, 'cost': round(v['cost'], 2), 'runs': v['runs']}
-                   for k, v in sorted(by_instance_map.items(), key=lambda x: -x[1]['cost'])]
-
-    # By dashboard
-    by_dash_map = {}
-    for run in runs_with_cost:
-        dash = run.get('dashboard', 'unknown')
-        if dash not in by_dash_map:
-            by_dash_map[dash] = {'cost': 0, 'runs': 0}
-        by_dash_map[dash]['cost'] += run['cost_usd']
-        by_dash_map[dash]['runs'] += 1
-    by_dashboard = [{'dashboard': k, 'cost': round(v['cost'], 2), 'runs': v['runs']}
-                    for k, v in sorted(by_dash_map.items(), key=lambda x: -x[1]['cost'])]
-
-    # Summary
-    total_cost = sum(run['cost_usd'] for run in runs_with_cost)
-    spot_cost = sum(run['cost_usd'] for run in runs_with_cost if run.get('spot'))
-
-    return Response(json.dumps({
-        'by_date': by_date,
-        'by_instance_type': by_instance,
-        'by_dashboard': by_dashboard,
-        'summary': {
-            'total_cost': round(total_cost, 2),
-            'spot_pct': round(100.0 * spot_cost / max(total_cost, 0.01), 1),
-            'avg_cost_per_run': round(total_cost / max(len(runs_with_cost), 1), 2),
-            'total_runs': len(runs_with_cost),
-        },
-    }), mimetype='application/json')
-
-
-# ---- CI Performance endpoint ----
-
-@app.route('/api/ci/performance')
-@auth.login_required
-def api_ci_performance():
-    date_from = request.args.get('from', (datetime.now() - timedelta(days=30)).strftime('%Y-%m-%d'))
-    date_to = request.args.get('to', datetime.now().strftime('%Y-%m-%d'))
-    dashboard = request.args.get('dashboard', '')
-    granularity = request.args.get('granularity', 'daily')
-    ts_from = int(datetime.strptime(date_from, '%Y-%m-%d').timestamp() * 1000)
-    ts_to = int((datetime.strptime(date_to, '%Y-%m-%d') + timedelta(days=1)).timestamp() * 1000)
-
-    runs = rk_metrics.get_ci_runs(r, ts_from, ts_to)
-    runs = [run for run in runs if run.get('status') in ('PASSED', 'FAILED')]
-    if dashboard:
-        runs = [run for run in runs if run.get('dashboard') == dashboard]
-
-    # By date
-    by_date_map = {}
-    for run in runs:
-        date = rk_metrics._ts_to_date(run.get('timestamp', 0))
-        if date not in by_date_map:
-            by_date_map[date] = {'total': 0, 'passed': 0, 'failed': 0, 'durations': []}
-        by_date_map[date]['total'] += 1
-        if run.get('status') == 'PASSED':
-            by_date_map[date]['passed'] += 1
-        else:
-            by_date_map[date]['failed'] += 1
-        complete = run.get('complete')
-        ts = run.get('timestamp')
-        if complete and ts:
-            by_date_map[date]['durations'].append((complete - ts) / 60000.0)
-
-    by_date = []
-    for date in sorted(by_date_map):
-        d = by_date_map[date]
-        by_date.append({
-            'date': date,
-            'total': d['total'],
-            'passed': d['passed'],
-            'failed': d['failed'],
-            'pass_rate': round(100.0 * d['passed'] / max(d['total'], 1), 1),
-            'failure_rate': round(100.0 * d['failed'] / max(d['total'], 1), 1),
-            'avg_duration_mins': round(sum(d['durations']) / len(d['durations']), 1) if d['durations'] else None,
-        })
-
-    by_date = _aggregate_dates(by_date, granularity,
-                               sum_fields=['total', 'passed', 'failed'],
-                               avg_fields=['avg_duration_mins'])
-    for d in by_date:
-        d['pass_rate'] = round(100.0 * d['passed'] / max(d['total'], 1), 1)
-        d['failure_rate'] = round(100.0 * d['failed'] / max(d['total'], 1), 1)
-
-    # Daily flake/failure counts from test_events
-    if dashboard:
-        flake_daily = rk_db.query('''
-            SELECT substr(timestamp, 1, 10) as date, COUNT(*) as count
-            FROM test_events WHERE status = 'flaked' AND dashboard = ?
-            AND timestamp >= ? AND timestamp < ?
-            GROUP BY substr(timestamp, 1, 10)
-        ''', (dashboard, date_from, date_to + 'T23:59:59'))
-        fail_test_daily = rk_db.query('''
-            SELECT substr(timestamp, 1, 10) as date, COUNT(*) as count
-            FROM test_events WHERE status = 'failed' AND dashboard = ?
-            AND timestamp >= ? AND timestamp < ?
-            GROUP BY substr(timestamp, 1, 10)
-        ''', (dashboard, date_from, date_to + 'T23:59:59'))
-    else:
-        flake_daily = rk_db.query('''
-            SELECT substr(timestamp, 1, 10) as date, COUNT(*) as count
-            FROM test_events WHERE status = 'flaked'
-            AND timestamp >= ? AND timestamp < ?
-            GROUP BY substr(timestamp, 1, 10)
-        ''', (date_from, date_to + 'T23:59:59'))
-        fail_test_daily = rk_db.query('''
-            SELECT substr(timestamp, 1, 10) as date, COUNT(*) as count
-            FROM test_events WHERE status = 'failed'
-            AND timestamp >= ? AND timestamp < ?
-            GROUP BY substr(timestamp, 1, 10)
-        ''', (date_from, date_to + 'T23:59:59'))
-    flake_daily_map = {r['date']: r['count'] for r in flake_daily}
-    fail_test_daily_map = {r['date']: r['count'] for r in fail_test_daily}
-    for d in by_date:
-        d['flake_count'] = flake_daily_map.get(d['date'], 0)
-        d['test_failure_count'] = fail_test_daily_map.get(d['date'], 0)
-
-    # Flake/failure data from test_events (the only SQLite data)
-    if dashboard:
-        top_flakes = rk_db.query('''
-            SELECT test_cmd, COUNT(*) as count, ref_name
-            FROM test_events WHERE status='flaked' AND dashboard = ?
-            AND timestamp >= ? AND timestamp <= ?
-            GROUP BY test_cmd ORDER BY count DESC LIMIT 15
-        ''', (dashboard, date_from, date_to + 'T23:59:59'))
-        top_failures = rk_db.query('''
-            SELECT test_cmd, COUNT(*) as count
-            FROM test_events WHERE status='failed' AND dashboard = ?
-            AND timestamp >= ? AND timestamp <= ?
-            GROUP BY test_cmd ORDER BY count DESC LIMIT 15
-        ''', (dashboard, date_from, date_to + 'T23:59:59'))
-    else:
-        top_flakes = rk_db.query('''
-            SELECT test_cmd, COUNT(*) as count, ref_name
-            FROM test_events WHERE status='flaked' AND timestamp >= ? AND timestamp <= ?
-            GROUP BY test_cmd ORDER BY count DESC LIMIT 15
-        ''', (date_from, date_to + 'T23:59:59'))
-        top_failures = rk_db.query('''
-            SELECT test_cmd, COUNT(*) as count
-            FROM test_events WHERE status='failed' AND timestamp >= ? AND timestamp <= ?
-            GROUP BY test_cmd ORDER BY count DESC LIMIT 15
-        ''', (date_from, date_to + 'T23:59:59'))
-
-    # Summary
-    total = len(runs)
-    passed = sum(1 for run in runs if run.get('status') == 'PASSED')
-    failed = total - passed
-    durations = []
-    for run in runs:
-        complete = run.get('complete')
-        ts = run.get('timestamp')
-        if complete and ts:
-            durations.append((complete - ts) / 60000.0)
-
-    if dashboard:
-        flake_count = rk_db.query('''
-            SELECT COUNT(*) as c FROM test_events WHERE status='flaked' AND dashboard = ?
-            AND timestamp >= ? AND timestamp <= ?
-        ''', (dashboard, date_from, date_to + 'T23:59:59'))
-        total_tests = rk_db.query('''
-            SELECT COUNT(*) as c FROM test_events WHERE status IN ('failed','flaked') AND dashboard = ?
-            AND timestamp >= ? AND timestamp <= ?
-        ''', (dashboard, date_from, date_to + 'T23:59:59'))
-    else:
-        flake_count = rk_db.query('''
-            SELECT COUNT(*) as c FROM test_events WHERE status='flaked' AND timestamp >= ? AND timestamp <= ?
-        ''', (date_from, date_to + 'T23:59:59'))
-        total_tests = rk_db.query('''
-            SELECT COUNT(*) as c FROM test_events WHERE status IN ('failed','flaked') AND timestamp >= ? AND timestamp <= ?
-        ''', (date_from, date_to + 'T23:59:59'))
-
-    fc = flake_count[0]['c'] if flake_count else 0
-    tc = total_tests[0]['c'] if total_tests else 0
-
-    if dashboard:
-        total_failures_count = rk_db.query('''
-            SELECT COUNT(*) as c FROM test_events WHERE status='failed' AND dashboard = ?
-            AND timestamp >= ? AND timestamp <= ?
-        ''', (dashboard, date_from, date_to + 'T23:59:59'))
-    else:
-        total_failures_count = rk_db.query('''
-            SELECT COUNT(*) as c FROM test_events WHERE status='failed' AND timestamp >= ? AND timestamp <= ?
-        ''', (date_from, date_to + 'T23:59:59'))
-    tfc = total_failures_count[0]['c'] if total_failures_count else 0
-
-    return Response(json.dumps({
-        'by_date': by_date,
-        'top_flakes': top_flakes,
-        'top_failures': top_failures,
-        'summary': {
-            'total_runs': total,
-            'pass_rate': round(100.0 * passed / max(total, 1), 1),
-            'failure_rate': round(100.0 * failed / max(total, 1), 1),
-            'avg_duration_mins': round(sum(durations) / len(durations), 1) if durations else None,
-            'flake_rate': round(100.0 * fc / max(tc, 1), 1) if tc else 0,
-            'total_flakes': fc,
-            'total_test_failures': tfc,
-        },
-    }), mimetype='application/json')
-
-
-# ---- Deployment speed endpoint ----
-
-@app.route('/api/deployments/speed')
-@auth.login_required
-def api_deploy_speed():
-    date_from = request.args.get('from', (datetime.now() - timedelta(days=30)).strftime('%Y-%m-%d'))
-    date_to = request.args.get('to', datetime.now().strftime('%Y-%m-%d'))
-    workflow = request.args.get('workflow', '')
-    granularity = request.args.get('granularity', 'daily')
-    result = rk_github.get_deployment_speed(date_from, date_to, workflow)
-    if granularity != 'daily' and result.get('by_date'):
-        result['by_date'] = _aggregate_dates(
-            result['by_date'], granularity,
-            sum_fields=['count', 'success', 'failure'],
-            avg_fields=['median_mins', 'p95_mins'])
-    return Response(json.dumps(result), mimetype='application/json')
-
-
-# ---- Branch lag endpoint ----
-
-@app.route('/api/branches/lag')
-@auth.login_required
-def api_branch_lag():
-    date_from = request.args.get('from', (datetime.now() - timedelta(days=30)).strftime('%Y-%m-%d'))
-    date_to = request.args.get('to', datetime.now().strftime('%Y-%m-%d'))
-    return Response(json.dumps(rk_github.get_branch_lag(date_from, date_to)), mimetype='application/json')
-
-
-# ---- PR metrics endpoint ----
-
-@app.route('/api/prs/metrics')
-@auth.login_required
-def api_pr_metrics():
-    date_from = request.args.get('from', (datetime.now() - timedelta(days=30)).strftime('%Y-%m-%d'))
-    date_to = request.args.get('to', datetime.now().strftime('%Y-%m-%d'))
-    author = request.args.get('author', '')
-    # Read CI runs from Redis to compute per-PR costs
-    ts_from = int(datetime.strptime(date_from, '%Y-%m-%d').timestamp() * 1000)
-    ts_to = int((datetime.strptime(date_to, '%Y-%m-%d') + timedelta(days=1)).timestamp() * 1000)
-    ci_runs = rk_metrics.get_ci_runs(r, ts_from, ts_to)
-    return Response(json.dumps(rk_github.get_pr_metrics(date_from, date_to, author, ci_runs)), mimetype='application/json')
-
-
-@app.route('/api/merge-queue/stats')
-@auth.login_required
-def api_merge_queue_stats():
-    date_from = request.args.get('from', (datetime.now() - timedelta(days=30)).strftime('%Y-%m-%d'))
-    date_to = request.args.get('to', datetime.now().strftime('%Y-%m-%d'))
-    return Response(json.dumps(rk_github.get_merge_queue_stats(date_from, date_to)), mimetype='application/json')
-
-
-@app.route('/api/ci/flakes-by-command')
-@auth.login_required
-def api_flakes_by_command():
-    """Flake counts grouped by CI command type (dashboard section)."""
-    date_from = request.args.get('from', (datetime.now() - timedelta(days=30)).strftime('%Y-%m-%d'))
-    date_to = request.args.get('to', datetime.now().strftime('%Y-%m-%d'))
-    dashboard = request.args.get('dashboard', '')
-    # Ensure test events are synced from Redis lists
-    rk_metrics.sync_failed_tests_to_sqlite(r)
-    return Response(json.dumps(rk_metrics.get_flakes_by_command(date_from, date_to, dashboard)), mimetype='application/json')
-
-
-# ---- Dashboard view routes ----
-
 @app.route('/ci-health')
-@auth.login_required
-def ci_health():
-    return redirect('/ci-insights')
-
-
 @app.route('/ci-insights')
-@auth.login_required
-def ci_insights():
-    path = Path('dashboard-views/ci-insights.html')
-    if path.exists():
-        return path.read_text()
-    return "Dashboard not found", 404
-
-
 @app.route('/cost-overview')
 @auth.login_required
-def cost_overview():
-    path = Path('dashboard-views/cost-overview.html')
-    if path.exists():
-        return path.read_text()
-    return "Dashboard not found", 404
+def proxy_dashboard():
+    try:
+        resp = requests.get(f'{CI_METRICS_URL}{request.path}', timeout=10)
+        return Response(resp.content, status=resp.status_code,
+                        content_type=resp.headers.get('Content-Type', 'text/html'))
+    except Exception as e:
+        return f"ci-metrics unavailable: {e}", 502
 
+@app.route('/api/<path:path>')
+@auth.login_required
+def proxy_api(path):
+    try:
+        resp = requests.get(f'{CI_METRICS_URL}/api/{path}', params=request.args, timeout=30)
+        return Response(resp.content, status=resp.status_code,
+                        mimetype=resp.headers.get('Content-Type', 'application/json'))
+    except Exception as e:
+        return Response(json.dumps({'error': f'ci-metrics unavailable: {e}'}),
+                        mimetype='application/json', status=502)
 
 @app.route('/<key>')
 @auth.login_required
